audio_gesture_mine/src/rwkv/src/rwkv/requirements.txt
opencv-python
mediapipe
numpy
python-osc
rtmidi
pyaudio
openmesh
scikit-learn
torch-scatter
